pyproject.toml
[project]
name = "chatkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.2",
    "jinja2>=3.1.6",
    "openai-agents>=0.8.0",
    "openai-chatkit>=1.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
